pyproject.toml
[project]
name = "teamchat"
version = "0.3.0"
description = "Default template for PDM package"
authors = [
    {name = "Nicou11", email = "lyoungin1204@gmail.com"},
]
dependencies = [
    "kafka-python>=2.0.2",
]
requires-python = "==3.8.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.3.2",
]
